Adicao de profiles

+++ app/Http/Controllers/Admin/PlanController.php
@@ -49,10 +49,15 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')->where('url', $url)->first();
         if (!$plan) {
             return redirect()->back();
         }
+
+        if($plan->details->count() > 0){
+            return redirect()->back()->with('error', 'Existem detalhes vinculados a esse plano, portanto, é necessário remover os detalhes primeiro!');
+        }
+
         $plan->delete();
 
         return redirect()->route('plans.index');
